+++ frontend/src/App.js
@@ -1,29 +1,26 @@
 import React from "react";
-import {BrowserRouter, Routes, Route} from "react-router-dom";
-import Home from "./Home";
-import Register from "./Register";
-import Login from "./Login";
-// import {Container} from "./component/Container";
-// import {Menu} from "./component/Menu";
-// import {StatusBar} from "./component/StatusBar";
 
-function App() {
+import { StatusBar } from "./components/layouts/StatusBar";
+import { Menu } from "./components/layouts/Menu";
+
+import { BrowserRouter, Routes, Route} from "react-router-dom";
+import { Home } from "./components/pages/home/Home";
+import { UserProfile } from "./components/pages/users/UserProfile";
+
+
+const App = () => {
     return(
-        // <div>
-        //     <StatusBar />
-        //     <Menu />
-        //     <Container />
-        // </div>
-        <BrowserRouter>
-            <Routes>
-                <Route path={`/`} element={<Home />} />
-                <Route path={`/register/`} element={<Register />}/>
-                <Route path={`/login/`} element={<Login />} />
-            </Routes>
-        </BrowserRouter>
+        <>
+            <BrowserRouter>
+                <Menu />
+                <StatusBar />
+                <Routes>
+                    <Route path="/" element={<Home />} />
+                    <Route path="user" element={<UserProfile />} />
+                </Routes>
+            </BrowserRouter>
+        </>
     );
 }
-
-
 
 export {App};
