--- frontend/src/component/Menu.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-class Menu extends React.Component {
-  render(){
-    return(
-      <div className="menu">
-        <h2>これはMenuです</h2>
-      </div>
-    );
-  }
-}
-
-export {Menu};
