--- frontend/src/component/StatusBar.js
@@ -1,14 +1,0 @@
-import React from "react";
-
-class StatusBar extends React.Component {
-  render(){
-    return(
-      <div className="status-bar">
-        <h2>これはStatusBarです</h2>
-      </div>
-    );
-  }
-
-}
-
-export { StatusBar };
