+++ frontend/src/components/layouts/Container.js
@@ -1,16 +1,12 @@
 import React from "react";
-import {UserShow} from "./contents/users/show";
 
-class Container extends React.Component {
-  render(){
+
+const Container = () => {
     return( 
       <div className="container">
         <h1>これはcontainerです</h1>
-        <UserShow />
       </div>
     );
-  } 
-
 }
 
 export { Container };
