+++ frontend/src/components/pages/home/Home.js
@@ -1,0 +1,13 @@
+import React from "react";
+import { Link } from "react-router-dom";
+
+const Home = () => {
+  return(
+    <div>
+      <h1>Home</h1>
+      <Link to="user">ユーザーページへ</Link>
+    </div>
+  );
+}
+
+export {Home};
