+++ frontend/src/index.js
@@ -3,6 +3,4 @@ import ReactDOM from "react-dom";
 import "./index.css"
 import {App} from "./App.js"
 
-
-
 ReactDOM.render(<App />, document.getElementById("root"));
